Initial gpupgrade commit
This commit comprises the initial development work on gpupgrade.  The project at
this point has all the necessary bare-bones functionality: it can be installed,
it can communicate with segments, it interacts correctly with the shell, it has
tests and documentation, and so forth.

Much of the code is heavily refactored or removed entirely in later commits,
but this commit provides the general framework for later development.

Co-authored-by: C.J. Jameson <cjameson@pivotal.io>
Co-authored-by: Chumki Roy <croy@pivotal.io>
Co-authored-by: Larry Hamel <lhamel@pivotal.io>
Co-authored-by: Marbin Tan <mtan@pivotal.io>
Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Tushar Dadlani <tdadlani@pivotal.io>