Migration and refactoring of functionality
This commit involves a lot of effort to make gpupgrade independent of other
code.  It modifies the Makefile to depend less on the rest of the GPDB build
process, copies in code previously imported from gpbackup, changes integration
test to build their own binaries instead of relying on pre-build binaries, and
so forth.

In addition to large amounts of normal refactoring, it also makes three major
architectural changes.  First, it switches inter-host communication from SSH
to gRPC; second, it introduces the Cobra framework for flag parsing and command
execution; finally, it adds a "hub" utility on the master and centralizes most
of the logic from the CLI and agent processes there.

Co-authored-by: C.J. Jameson <cjameson@pivotal.io>
Co-authored-by: Larry Hamel <lhamel@pivotal.io>
Co-authored-by: Marbin Tan <mtan@pivotal.io>
Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Tushar Dadlani <tdadlani@pivotal.io>