Standardization of state management
This commit introduces the handling of cluster state and upgrade status using
lightweight state files, and modifies most gpupgrade commands to standardize on
that method of tracking state.

It also introduces and modifies various commands that depend on knowing the
current state of the cluster or the current upgrade status, and makes
improvements around logging and error reporting.

Finally, this commit introduces and standardizes testing patterns for testing
asynchronous functions, working with end-to-end tests, and so forth.

Co-authored-by: C.J. Jameson <cjameson@pivotal.io>
Co-authored-by: Marbin Tan <mtan@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>