Migrate gpupgrade to its own repo
This commit is the first made after migrating gpupgrade from a subdirectory of
the main GPDB repo to its own repository.  As part of the migration, the README
and Makefile were updated to account for gpupgrade no longer being built along
with GPDB, and the version number was changed from using the GPDB version to
using an independent gpupgrade version number based on the current build number
and latest tag.

Additionally, gpupgrade was renamed to "gpupgrade" from its previous name of
"gp_upgrade" to comport with the standard naming convention of the Greenplum
utilities.  This includes all references to "gp_upgrade" in filenames, import
paths, shelled-out commands, make targets, and documentation.

Authored-by: Jamie McAtamney <jmcatamney@pivotal.io>