StateCheck.GetStatus: don't return an error
The error condition for GetStatus() is not particularly well-defined.
All callers ignored the error entirely, except for in the unit tests,
which meant that we would error during gRPC marshalling anyway...

Instead of pretending like we care about an error that we ignore, log
any errors and mark the step as PENDING. We may want to revert this
decision later as the steps get more complicated and more difficult to
"just run again", but when we do that we should thread proper error
handling through all callers.

Co-authored-by: Nadeem Ghani <nghani@pivotal.io>