Fill out the Checklist interface with StateReader
ChecklistWriter is now just Checklist; it encapsulates both the reading
and the writing sides of state management. StateCheck now implements
the StateReader interface, so it can be returned directly when clients
ask for a reader from the Checklist; nothing else needs to actually know
about StateCheck.

Many tests now use the MockChecklistManager out of necessity. It's
becoming obvious that some tests will need to be changed in scope or
removed entirely.

Co-authored-by: Jim Doty <jdoty@pivotal.io>