Pull `status upgrade` implementation all the way to the top
Move all business logic for the status upgrade steps to the hub's main
package. To make this possible:

Checklist is now in charge of storing the list of steps; they are loaded
during hub initialization. This moves most of the business logic out of
the `status upgrade` implementation; it is able to simply query the
hub's checklist for a series of steps to call in order.

Since the ProtoBuf code for each step is also stored in the Checklist,
status implementations don't need to pass around UpgradeStepStatus (a
source of massive duplication in the code). They all return StepStatus
directly.

The master-upgrade unit tests in hub_status_upgrade_test are no longer
useful, since they now test the mock implementation.
pg_upgrade_status_checker_test contains a unit test for each case, and
the superfluous tests can be safely removed.

Two integration tests have been broken and X'd out for now; it's not
clear whether they add enough testing value at the moment to fix.
Current leaning is that the pieces they test should be moved into an
end-to-end suite.

Hooray for mobs!

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>