CLI: give a useful error message if the hub isn't up
gRPC uses asynchronous connections by default. This is an excellent
default for a high-performance, multi-connection server/client system,
but it doesn't make as much sense for our client since we can't do any
useful work in parallel. It also means that a failure to connect is
treated as a "transient" failure, which leads to the less-than-helpful

    rpc error: code = Unavailable desc = all SubConns are in TransientFailure

error message if you've forgotten to start the hub.

Switch to blocking connections for the client. The default connection
timeout is one second, which can be modified using the
GPUPGRADE_CONNECTION_TIMEOUT environment variable. The new test uses
this variable so that it doesn't take forever.