Hub: always close connections during Stop()
The "Hub closes open connections when shutting down" test failed
intermittently because we did not close open connections if the h.server
member didn't exist. Hub.AgentConns() doesn't actually use h.server, nor
does Hub.closeConns(), and it is possible to call Hub.AgentConns()
without the server having been started. So with the following code:

    go hub.Start()
    conns := hub.AgentConns()
    hub.Stop()

it is possible for hub.Stop() to be invoked before hub.Start(), in which
case the connections that we just opened will not be closed.

To fix, always close the connections. It is valid to range over a nil
slice, so it's okay to call h.closeConns() on a server that doesn't have
connections open yet -- but we do need to make sure that accesses to
h.agentConns are synchronized; hence the addition of the mutex lock in
AgentConns().

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>