Hub: fix race between Start() and Stop()
Currently,

    go hub.Start()
    hub.Stop()

is not actually guaranteed to end with a stopped Hub, because Start()
doesn't make an explicit check to see if it's already been stopped. This
commit fixes that by nil'ing out the stopped channel during Stop() so
that Start() can't use it.

The new test is subject to a transient failure: if Stop() is called
after Start() creates a gRPC server but before that server is started,
the server will return an error from server.Serve() and we will panic.
This will be fixed in a follow-up commit.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>