make: don't install during build
Previously we were putting binaries into the GOPATH during build. But
the user should really be able to run `make check` before installing, so
that we don't accidentally install a build that doesn't pass the tests.

Instead, put the binaries into the current directory, and make use of
them during the end-to-end tests that are run as part of `make check`.
They will still be installed under the GPDB bin/ directory during
installation. (Note that the integration tests build their own versions
of the utilities already and don't need to be modified here.)

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>