Split ClusterPair into two Clusters
Having a ClusterPair containing two cluster.Cluster structs, rather than putting
two cluster.Clusters on the hub, did not give us any functional benefit, except
for the fact that we could not add additional helper functions onto a struct
from a different package.

This commit introduces a utils.Cluster that wraps a cluster.Cluster, allowing us
to add helper functions and to store the binary directory and config file path
in each cluster instead of a ClusterPair.  It correspondingly refactors the hub
functions to use two utils.Clusters instead of a ClusterPair.

Additionally, because referencing "old" and "new" clusters has caused confusion
in the past, this commit changes the convention to use "source" and "target"
instead, to match the use of "target cluster" in prepare init-cluster.

As of this commit, the integration tests are working, but the unit tests require
additional attention.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>