make: deploy-pipeline is git aware
Previously the make target 'deploy-pipeline' was not savy to the
workflow differences of iterating on a branch as compared to setting
master. This commit removes this friction.  Now `make deploy-pipeline`
will try and do the right thing based on your git status.  If you are
checked out on master and tracking the production fork of the repo, it
should set the production version of the pipeline.  If not it will
interact with the dev deployment of our CI system.

Authored-by: Jim Doty <jdoty@pivotal.io>