ci: use resource mapping in pipeline
It can be helpful as a pipeline grows to have the names of the resources
contain a hint to the contents of the resource, for example gpdb_src is
distiguisable from bin_gpdb in the master GPDB pipeline. Therefore this
commit suggests that we use `gpupgrade_src` in place of `gpupgrade`.
This being said, the nature of the golang toolchain requires that the
cloned repo live in `$GOPATH/src/.../gpupgrade`. We can have both if we
use the resource attribute in the `get: gpupgrade` step.