ci: try using the built in go env
The container image that we are using already has a built in go
installation and go environment setup.  This commit reduces the need to
set up the environment in the task yml, however in order to place our
code in that env, we have to use a relative path `../../../go` to get
to the built in `/go`.  This is because Concourse specifically ignores
absolute paths in task definitions[1].

[1] https://concourse-ci.org/tasks.html#input-path

Authored-by: Jim Doty <jdoty@pivotal.io>