testutils: change multinode cluster to actually be multinode
The multinode cluster didn't actually use multiple nodes. This commit
rectifies that situation, and updates many tests to not assume that the
only host under test is localhost.

NewMockAgentServer() now provides a mock services.Dialer that will
always redirect connections to the mock server, no matter what host is
passed. This prevents tests from failing simply because they're trying
to connect to some host other than localhost.

The Hub needs to only generate an Agent connection for hosts where a
primary is running; the master won't have an agent. This affects the
implementation of, and several tests for, the primary conversion logic,
since master is no longer "special" to the agents. utils.Cluster has
been given a PrimaryHostnames() helper to facilitate this.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>