seginstall: error out if no segments are configured
`check seginstall` was silently "succeeding", even after having done no
work, when the segment configuration was blank. This could happen right
after a new hub is started but before `check config` is run.

Fix this by using the new Cluster.ExecuteOnAllHosts() helper, which
returns an error if the cluster has no segments. A quick refactor to
VerifyAgentsInstalled() lets us handle this error at the top level of
the CheckSeginstall() goroutine.

Co-authored-by: Jim Doty <jdoty@pivotal.io>