Rename convert-master and convert-primaries status directories
When fixing gpupgrade status upgrade we found that the reason running
convert-master would also update the status for convert-primaries when
running status upgrade was that both steps called their status directory
"pg_upgrade" and so checking the status for one step would look in the
other step's directory as well.

There is no need to name the directory "pg_upgrade" instead of the step
name, like the other step status directories, so this commit renames
them to avoid the namespace collision.

As part of this commit, we also changed status conversion to no longer
show the master status, as the convert master status line in status
upgrade is sufficient to cover that.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>