Run pg_upgrade as child process of the hub
Previously when the hub serviced the ConvertMaster request, it would
attempt to nohup the pg_upgrade process.  The way this was implemented
meant that we had difficulty knowing when the proccess failed.  We
changed the pattern and "merely" execute the ConvertMaster() function
from a goroutine, and have that function return an error.  This makes
the CONVERT_MASTER step writeable.

Notes:

This change leave the agents as the last users of the upgradestatus
checking logic, so if we decide to pull pg_upgrade into goroutines on
the agents, we will be able to scrub that logic out of the code base.

We also observed that the pg_upgrade flag --dispatcher-mode had changed
to --mode=dispatcher, and pg_upgrade also failed to start if passed the
--progress flag (observed at gpdb/master commit 255c3c1e37).

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>