utils.Cluster: add Version member
We need to have some idea of what cluster version we're dealing with,
even after both the source and target clusters have been shut down, so
that we know what data to migrate between the master and primaries after
the master upgrade. Add Cluster.Version and populate it during
utils.ClusterFromDB().

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>