CI: install BATS
To enable `make check` in a future commit, pull down bats-core as a
Concourse resource and install it using the provided helper script. Lock
the BATS version to a released 1.x tag to avoid any master instability.

Note: on the golang container, there is no TERM set, and BATS needs that
envvar to provide pretty-printing. This seems like a potential bug that
could be reported, but set the TERM manually for now.