convert-*: source greenplum_path.sh before upgrading
The GPDB build of pg_upgrade is linked against libraries in the GPDB
installation path, so it (unfortunately) needs the LD_LIBRARY_PATH
modifications in greenplum_path.sh. This is true for both the hub and
agent invocations of pg_upgrade, but it's more likely to show up as a
problem for the agent, since the hub generally inherits the user's
environment on start. (This, by the way, seems like a potential pitfall
in the future...)

One day we'll use rpath like we should, but until then, source the
environment script before running pg_upgrade.