Modify status order and sort at CLI
Previously, the ordering of the gpupgrade status upgrade entries was fixed based
on the order they were added to the hub; changing the display order thus
required recompiling and restarting the hub.

This commit changes the logic to sort the status messages at the CLI, enabling
the order to be changes without restarting the hub.  It also modifies the order
of the steps in the hub to be closer to the intended upgrade workflow so the
order is more obvious when looking at that code.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>