convert-master: stream stdout/err from pg_upgrade
...using exec.Command directly so that we can get at the independent
streams. This effectively makes convert-master synchronous. Client
disconnections will not affect pg_upgrade execution; it will continue
and all output will still be streamed to a log file.

Also reimplement the core of ConvertMaster() to be fully unit-testable.
There are still end-to-end tests that we'd like to write for this case.

Due to a bug in mockgen, we're unable to use the "source mode"
generation when using gRPC streams. We have to manually specify the mock
interfaces we want to generate, using "reflect mode"; hence the changes
to idl/generate.go.

Co-authored-by: Mark Sliva <msliva@pivotal.io>