exectest: don't use exec.Cmd.Env for magic
We were using the Cmd environment to pass information to exectest.Run()
inside the test subprocess, but that conflicts with legitimate uses of
Cmd.Env by the actual implementation. Technically any use of the Cmd
struct to contain this out-of-band data could conflict with legitimate
use cases, but for now we kick the can down the road and pass the magic
index using cmd.Args[0] instead. This would fail if any implementations
explicitly set cmd.Args[0], but we don't currently forsee any such
implementations in gpupgrade. (Famous last words.)

As a side effect, the original arguments passed to exec.Command are
faithfully reproduced in the proper location in the subprocess;
previously, they were shifted to the right by one, so that os.Args[1:]
was the "actual" argument list.

Co-authored-by: Mark Sliva <msliva@pivotal.io>