exectest: allow multiple calls to RegisterMains()
It didn't occur to me that, for exectest to be usable from multiple test
files in the same package, we need to allow multiple calls to
RegisterMains(). Because Go defines a static order for init() execution,
determined at compile time, this is safe so long as RegisterMains() is
only called from init().