make check: don't implicitly rely on installation paths
The install-tests pipeline job correctly catches our mistake: by passing
"${GPHOME}/bin" as the --new-bindir, we are testing the installed
gpupgrade_agent executable instead of the locally-built one. Fix this by
switching to $PWD...

...which breaks the gpinitsystem test. Until we fix our executable
lookup problem, by expecting the gpupgrade binaries to be generally
available on the PATH, we won't be able to run that test correctly.

Co-authored-by: David Krieger <dkrieger@pivotal.io>