Assume agent path is the same as the hub path
We assume that the gpupgrade binaries are installed in the same location
on all the cluster hosts. Therefore, the full path of the agent
executable is derived from the path of the hub.

For CI upgrade tests copying gpupgrade binaries into /usr/local/bin
which is on the PATH.

With this commit and having gpupgrade in the PATH, we no longer skip
execute.bats.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>
Co-authored-by: Mark Sliva <msliva@pivotal.io>