initialize: check available disk space
Replace the old disk space checks with disk.CheckUsage(). This
implementation is run by both the hub and the agents during initialize;
it returns a list of any filesystems that do not meet the requirements.
(For now, only data directory roots are checked; tablespaces are left
for later.)

Since we only support copy-mode for now, the default target disk
availability is 60%. This can be overridden via the --disk-free-ratio
option to initialize.

Object-count checks were unused and have been removed.

Co-authored-by: Mark Sliva <msliva@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>