hub/services_test: skip three-second connection timeouts
Unfortunately, the error returned by errors.New() from our dialer
implementation was treated as "temporary" by gRPC, so the dialer was
invoked over and over again up until the connection timeout. This
bloated the services tests by an additional six seconds of runtime.

Explicitly mark the errors as final, by implementing the Temporary()
interface. (IMNSHO, this is a bug in gRPC.)