upgrade-cluster.bash: factor helpers up out of SSH
At some point, functions that were meant to be run on the CCP container
host got moved to running on mdw, but the implementations still SSH'd to
mdw in a loopback. This makes our work on diff filtering harder, so undo
that change.

As a bonus, the implementation is much simpler now, since complex code
is no longer hiding inside a heredoc.

Co-authored-by: Mark Sliva <msliva@pivotal.io>