CI: filter/patch to fix the 5-to-6 pipeline diff
Unfortunately, there are some diffs in the output of pg_dumpall after
an upgrade from 5X to 6X, and we're probably not going to be able to
change them:

- The plpgsql extension is installed by default in 6, which leads to a
  group of related SQL additions.
- The postgres database now has an attached comment due to a change in
  behavior in the 6X initdb.
- All databases will be dumped with gp_use_legacy_hashops = 'on' in the
  6X cluster, but there has been a product decision not to dump the same
  setting for a 5X cluster.

To mask these changes, implement a two-step process. First use a filter
executable to remove any changes that can be handled with a simple
sed-like streaming approach. This still leaves diffs that are hard to
handle with a stream parser, so patch out those with a record of the
"acceptable diffs" before performing the final comparison.

Co-authored-by: Mark Sliva <msliva@pivotal.io>