CI: reenable installcheck script
installcheck isn't nearly as useful as it was originally, but the fact
that the CI wasn't running it at all is a little bit more disturbing.
Switch install-tests from a stdin heredoc, which was being silently
consumed by something inside `make check`, to a simple command string.
(This has the side effect of reenabling pretty-print for the BATS suite,
so we must export a TERM setting just like noinstall-tests does.)

Now that installcheck is running, fix up the bats script to handle
recent-ish changes to gpupgrade and the GPDB environment.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>