hub/agent: remove services package
The hub and agent package names are very nice, but we're currently not
using them at all. Get rid of the less-useful services layer.

To avoid combining too many code motions and changes all at once, I've
kept the "services" package aliases in this commit; they'll be removed
in the followups.