Persist the target master database port
In order to achieve idempotency, the master data port should be
persisted after the configuration of the new cluster is created. This
allows passing the master port to other steps in case of re-running of
the failed step.

Earlier after completion of CREATE_TARGET_CONFIG step the value of
target master database port was available, however if the next step
INIT_TARGET_CLUSTER failed and was re-run, the CREATE_TARGET_CONFIG will
not be executed and the value of master port will be unknown. Thus,
persist the value of master database port.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>