Remove the segment and master data directories if existing
If INIT_TARGET_CLUSTER step failed, while it's re-run the master and
segment data directories must be removed before proceeding to initialize
the cluster. In order to identify if the data directories are created by
a previous run of gpupgrade initialize, a marker file .gpupgrade is
created in the data directory. During re-run of the step, if the marker
file exists, the segment data directories are recreated, else an error
is raised requesting manual removal of the data directories.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>