pg_upgrade --check: add --verbose to debug test flakes
The test

    pg_upgrade --check fails on a source cluster with an index on a partition table

has been failing intermittently in the CI. Add --verbose output to try
to help catch the root cause, and don't use run() so that the output
gets flushed to the CI on a failure.

While we're at it, remove another unhelpful use of run() in the same
file.