Make start agents idempotent
To make the start agents substep idempotent we reuse the RestartAgents
implementation. We will need to revisit starting and stopping of the hub
and agents as our current implementation has some drawbacks. Namely, we
have no way of stopping spurious agents that the hub is not aware of.
PID files seems like the way to go.

This introduces a test helper program port_listener/main.go to listen on
all ipv4 and ipv6 interfaces on a specific port. If no ipv6 interface is
found it will not listen on ipv6. This is useful where our current CI
infrastructure does not support ipv6.

We use port_listener/main.go instead of netcat since this guarantees to
bind to all interfaces, and keeps the held gRPC connection open during
gRPC retries.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>